+++ src/lib/WP3TabGroup.h
@@ -40,7 +40,7 @@ public:
  private:
  	uint8_t m_subGroup;
 	uint8_t m_modeType;
-	float m_position;
+	double m_position;
 };
 
 #endif /* WP3TABGROUP_H */
